requirements.txt
# Core — install PyTorch with the CUDA 12.8 index:
#   pip install torch --index-url https://download.pytorch.org/whl/cu128
torch>=2.5.0

# Pinned for qwen-tts compatibility (qwen-tts requires transformers==4.57.3)
transformers==4.57.3

# HuggingFace Hub — model download & caching
huggingface_hub>=0.24.0

# Required by transformers device_map= support
accelerate>=0.30.0

# CUDA extension build (kernel.cu → .so)
ninja>=1.11.0

# Audio output for TTS integration (optional, not needed for parity check)
soundfile>=0.12.0
qwen-tts>=0.1.0
pipecat-ai>=0.0.1
websockets>=12.0  # required by pipecat at runtime
loguru>=0.7.0  # used by demo_pipecat.py
